NavBar changes, Undo functions, Superior shift view functions
NavBar should now be attached to the top of root in Main
CSS styles for Undo button
New method notifyUndo to ViewModule inteface, this is to help out with
undo functions.
New method to IView, used to reset the bottom bar. Right now Undo
message goes to the bottom part of root where Logout and language
buttons are. Use this to reset Logout and language buttons back to the
bottom after dismissing/autoclosing Undo message.